Rename Angelscript integration header to nvgt_angelscript.h. Also rename the implementation file for parody.
