+++ src/nvgt_angelscript.h
